Split input_mode
